+++ test/dummy/app/javascript/application.js
@@ -3,5 +3,6 @@ import { DemoCounterComponent } from "demo_counter_component";
 
 const achilles = new Application();
 achilles.componentsClassMapper.addComponentClass("DemoCounterComponent", DemoCounterComponent);
+achilles.start();
 
 window.achilles = achilles;